Collision bug solved
The collision with pipe would occur even at with a gap inbetween the player and the pipe. The issue was that the pipe width was taken as limit between top left corner of player and top left corner of pipe.
Changed the minimum distance to avoid collision to player width solving the bug.