High score display and pipe distance adjustment
The welcome screen now displays the highest score of all the games played in a single run of the program.
The high score will reset on restarting the program as no persistent storage is implemented as of now.

The 2nd and 3rd pipes were spawning too near making a blocked path if gaps were not aligned.
Removed generating 2 pipes at the start and only created a single pipe at the beginning of the game to solve the issue.